Refactor Redis integration and enhance error handling in file processing; remove unused query processing worker.

+++ queues/file_to_db_worker.py
@@ -10,33 +10,39 @@ load_dotenv()
 
 
 def file_embedding_and_loading(file_path, collection_name="ragpdf"):
-    
-    loader = PyMuPDFLoader(file_path=file_path)
-    docs = loader.load()
+    try:
+        loader = PyMuPDFLoader(file_path=file_path)
+        docs = loader.load()
 
-    text_splitter = RecursiveCharacterTextSplitter(
-        chunk_size=1000,
-        chunk_overlap=400
-    )
-    chunk = text_splitter.split_documents(docs)
+        text_splitter = RecursiveCharacterTextSplitter(
+            chunk_size=1000,
+            chunk_overlap=400
+        )
+        chunk = text_splitter.split_documents(docs)
 
-    embedding_model= GoogleGenerativeAIEmbeddings(
-        model="models/gemini-embedding-001"
-    )
+        embedding_model= GoogleGenerativeAIEmbeddings(
+            model="models/gemini-embedding-001"
+        )
 
-    QDRANT_URL = os.getenv("QDRANT_URL")
-    QDRANT_API = os.getenv("QDRANT_API")
+        QDRANT_URL = os.getenv("QDRANT_URL")
+        QDRANT_API = os.getenv("QDRANT_API")
 
-    vector_db = QdrantVectorStore.from_documents(
-        documents=chunk,
-        embedding=embedding_model,
-        url=QDRANT_URL,
-        api_key=QDRANT_API,
-        collection_name=collection_name
-    )
+        vector_db = QdrantVectorStore.from_documents(
+            documents=chunk,
+            embedding=embedding_model,
+            url=QDRANT_URL,
+            api_key=QDRANT_API,
+            collection_name=collection_name
+        )
 
-    print(f"Processed and stored {len(chunk)} chunks from {file_path} in Qdrant.")
+        print(f"Processed and stored {len(chunk)} chunks from {file_path} in Qdrant.")
 
-    os.remove(file_path)
+        os.remove(file_path)
 
-    return "File processed and indexed successfully."
+        return "File processed and indexed successfully."
+    except Exception as e:
+        print(f"Error processing file {file_path}: {e}")
+        return f"Error processing file: {e}"
+    finally:
+        if os.path.exists(file_path):
+            os.remove(file_path)
